fix: App page tile icons + More Apps launch behavior (LC-1928)
Two bugs from the App page (launchpad) redesign:

- LearnCard app icons rendered too small at tablet / in-between widths.
  The grid stayed 3-col until md (768px) while the tiles were capped at
  max-w-[100px], so at ~400–767px the tiles sat tiny inside much wider
  grid columns. Cap the tile at max-w-[160px] at every width (not just
  desktop) and shift the 3->4 column breakpoint from md to sm, so the
  original p-[18%] icon padding stays proportional across all viewports.

- More Apps tiles always opened the app detail modal. Installed apps now
  launch on tap — mirroring the "Open" button on the browse list — while
  not-installed (suggested) apps still open the detail page so the user
  can install first.

To avoid a third copy of the launch flow, the shared logic (consent
redirect + delegate VP, AI Tutor, embedded iframe, direct/second-screen
links, age hard-block, fallback) is extracted from AppStoreListItem into
a useAppLaunch hook now consumed by both AppStoreListItem and the new
MoreAppTile.
