+++ apps/learn-card-app/src/pages/myApps/AppGrid.tsx
@@ -14,7 +14,7 @@ const AppGrid: React.FC<AppGridProps> = ({ heading, children }) => (
         )}
         <div
             data-testid="app-grid"
-            className="grid grid-cols-3 items-start justify-items-center gap-x-3 gap-y-5 md:grid-cols-4 md:gap-x-6 md:gap-y-8"
+            className="grid grid-cols-3 items-start justify-items-center gap-x-3 gap-y-5 sm:grid-cols-4 md:gap-x-6 md:gap-y-8"
         >
             {children}
         </div>
