+++ apps/learn-card-app/src/pages/myApps/MyAppsLanding.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 import { IonPage, IonContent } from '@ionic/react';
-import { useModal, ModalTypes, useDeviceTypeByWidth } from 'learn-card-base';
+import { useDeviceTypeByWidth } from 'learn-card-base';
 
 import Search from 'learn-card-base/svgs/Search';
 
 import MainHeader from '../../components/main-header/MainHeader';
 import ProfileAlertsIsland from '../../components/main-header/ProfileAlertsIsland';
-import AppStoreDetailModal from '../launchPad/AppStoreDetailModal';
 import AppGrid from './AppGrid';
 import AppGridTile from './AppGridTile';
+import MoreAppTile from './MoreAppTile';
 import {
     LEARNCARD_APP_SHORTCUTS,
     JOURNEYS_SHORTCUT,
@@ -35,8 +35,6 @@ const MOBILE_HEADER_STYLE: React.CSSProperties = {
 const MyAppsLanding: React.FC = () => {
     const history = useHistory();
     const { search } = useLocation();
-    // Initialize with a modal type so newModal renders (mirrors AppStoreListItem).
-    const { newModal } = useModal({ desktop: ModalTypes.Right, mobile: ModalTypes.Right });
     const { isMobile } = useDeviceTypeByWidth();
     const openBoost = useOpenBoostTemplateSelector();
     const { apps: moreApps, isSuggested, isLoading: isLoadingMore } = useMoreApps();
@@ -99,14 +97,7 @@ const MyAppsLanding: React.FC = () => {
     );
 
     const renderAppTile = (app: (typeof moreApps)[number]) => (
-        <AppGridTile
-            key={app.listing_id}
-            title={app.display_name}
-            icon={app.icon_url}
-            onClick={() =>
-                newModal(<AppStoreDetailModal listing={app} isInstalled={!isSuggested} />)
-            }
-        />
+        <MoreAppTile key={app.listing_id} listing={app} isInstalled={!isSuggested} />
     );
 
     const renderSkeletonTile = (key: string) => (
